Added gui and gui main
